+++ src/libromdata/RomDataFactory.cpp
@@ -288,7 +288,7 @@ const RomDataFactoryPrivate::RomDataFns RomDataFactoryPrivate::romDataFns_header
 	GetRomDataFns(MegaDrive, ATTR_SUPPORTS_DEVICES),	// ATTR_SUPPORTS_DEVICES for Sega CD
 	GetRomDataFns(N64, ATTR_NONE | ATTR_HAS_METADATA),
 	GetRomDataFns(NES, ATTR_NONE),
-	GetRomDataFns(SNES, ATTR_NONE),
+	GetRomDataFns(SNES, ATTR_HAS_THUMBNAIL),
 	GetRomDataFns(SegaSaturn, ATTR_NONE | ATTR_HAS_METADATA | ATTR_SUPPORTS_DEVICES),
 	GetRomDataFns(WiiSave, ATTR_HAS_THUMBNAIL),
 	GetRomDataFns(WiiU, ATTR_HAS_THUMBNAIL | ATTR_SUPPORTS_DEVICES),