+++ src/rp-download/rp-download.cpp
@@ -409,6 +409,7 @@ int RP_C_API _tmain(int argc, TCHAR *argv[])
 	// - amiibo: https://amiibo.life/[key]/image
 	// - gba:    https://rpdb.gerbilsoft.com/gba/[key]
 	// - gb:     https://rpdb.gerbilsoft.com/gb/[key]
+	// - snes:   https://rpdb.gerbilsoft.com/snes/[key]
 	const TCHAR *slash_pos = _tcschr(cache_key, _T('/'));
 	if (slash_pos == nullptr || slash_pos == cache_key ||
 		slash_pos[1] == '\0')
@@ -467,8 +468,9 @@ int RP_C_API _tmain(int argc, TCHAR *argv[])
 			_T("https://amiibo.life/nfc/%.*s/image"),
 			static_cast<int>(filename_len), slash_pos+1);
 	} else if ((prefix_len == 3 && !_tcsncmp(cache_key, _T("gba"), 3)) ||
-		   (prefix_len == 2 && !_tcsncmp(cache_key, _T("gb"), 2))) {
-		// Game Boy, Game Boy Color, Game Boy Advance
+		   (prefix_len == 2 && !_tcsncmp(cache_key, _T("gb"), 2)) ||
+		   (prefix_len == 4 && !_tcsncmp(cache_key, _T("snes"), 4))) {
+		// Game Boy, Game Boy Color, Game Boy Advance, Super NES
 		_sntprintf(full_url, _countof(full_url),
 			_T("https://rpdb.gerbilsoft.com/%s"), cache_key);
 	} else {